Litigator final claim API implementation.
