+++ app/interfaces/api/v1/external_users/claims/advocate_claim.rb
@@ -6,9 +6,10 @@ module API::V1::ExternalUsers
       params do
         use :common_params
 
-        optional :advocate_email, type: String, desc: "REQUIRED: The ADP account email address that uniquely identifies the advocate to whom this claim belongs."
+        optional :advocate_email, type: String, desc: 'DEPRECATED: Use instead user_email.'
+        optional :user_email, type: String, desc: 'REQUIRED: The ADP account email address that uniquely identifies the advocate to whom this claim belongs.'
+
         optional :advocate_category, type: String, desc: "REQUIRED: The category of the advocate", values: Settings.advocate_categories
-        optional :offence_id, type: Integer, desc: "REQUIRED: The unique identifier for this offence"
         optional :first_day_of_trial, type: String, desc: "REQUIRED: YYYY-MM-DD", standard_json_format: true
         optional :estimated_trial_length, type: Integer, desc: "REQUIRED: The estimated trial length in days"
         optional :actual_trial_length, type: Integer, desc: "REQUIRED: The actual trial length in days"
